Tidying up new functions to get calculated summary pages
