+++ eq-author/src/App/page/Design/CalculatedSummaryPageEditor/AnswerSelector/index.js
@@ -12,6 +12,7 @@ import AnswerPicker from "components/AnswerPicker";
 
 import { useQuestionnaire } from "components/QuestionnaireContext";
 import getContentBeforeEntity from "utils/getContentBeforeEntity";
+import getCalculatedSummaryPages from "utils/getCalculatedSummaryPages";
 
 const Container = styled.div`
   border: 1px solid ${colors.grey};
@@ -23,74 +24,13 @@ const Container = styled.div`
 export const filterAvailableAnswers = (entity) =>
   [CURRENCY, UNIT, PERCENTAGE, NUMBER].includes(entity.type) ? entity : [];
 
-const insertSection = (sections, section, folder, page, answers) => {
-  sections.push({
-    ...section,
-    folders: [],
-  });
-  sections = insertFolder(sections, folder, page, answers);
-  return sections;
-};
-
-const insertFolder = (sections, folder, page, answers) => {
-  sections[sections.length - 1].folders.push({
-    ...folder,
-    pages: [],
-  });
-  sections = insertPage(sections, page, answers);
-  return sections;
-};
-
-const insertPage = (sections, page, answers) => {
-  sections[sections.length - 1].folders[
-    sections[sections.length - 1].folders.length - 1
-  ].pages.push({
-    ...page,
-    answers,
-  });
-  return sections;
-};
-
 const AnswerSelector = ({ page, onUpdateCalculatedSummaryPage }) => {
   const [showPicker, setShowPicker] = useState(false);
 
   const { questionnaire } = useQuestionnaire();
 
-  const getAllCalculatedSummaryPages = (pageId, sectionId) => {
-    const sections = [];
-
-    for (const section of questionnaire.sections) {
-      if (section.id === sectionId) {
-        return sections;
-      }
-      for (const folder of section.folders) {
-        for (const page of folder.pages) {
-          const answers = page?.answers;
-          if (page.pageType === "CalculatedSummaryPage") {
-            if (page.id === pageId) {
-              return sections;
-            }
-            if (sections[sections.length - 1]?.id === section?.id) {
-              if (
-                sections[sections.length - 1].folders[
-                  sections[sections.length - 1].folders.length - 1
-                ]?.id === folder?.id
-              ) {
-                insertPage(sections, page, answers);
-              } else {
-                insertFolder(sections, folder, page, answers);
-              }
-            } else {
-              insertSection(sections, section, folder, page, answers);
-            }
-          }
-        }
-      }
-    }
-    return sections;
-  };
-
-  const calculatedSummaries = getAllCalculatedSummaryPages(
+  const calculatedSummaries = getCalculatedSummaryPages(
+    questionnaire,
     page.id,
     page.section.id
   );
